Use a `Searcher` singleton across all searches to improve performance
In the current implementation, every search opens/loads Whoosh index and initiates a new `Searcher` instance. This commit introduces a singleton to eliminate such behaviour, except when syncing  attachments that requires the new index. The downside of this change is that after syncing, the Flask app needs to be restarted for the searcher to use the new index.